refactor: move deposits and withdrawals logic from sdk to ui (#637)
