+++ .github/workflows/build-test.yml
@@ -54,7 +54,7 @@ jobs:
         uses: ./.github/actions/build-artifacts/restore
 
       - name: Start UI and Test
-        run: yarn start-test 'dev:ui' http://localhost:3000 'yarn test:ci:sdk'
+        run: yarn start-server-and-test 'dev:ui' http://localhost:3000 'yarn test:ci:sdk'
 
   test-ui:
     name: "Test UI"
@@ -70,8 +70,8 @@ jobs:
       - name: Restore build artifacts
         uses: ./.github/actions/build-artifacts/restore
 
-      - name: Test
-        run: yarn test:ci:ui
+      - name: Start UI and Test
+        run: yarn start-server-and-test 'dev:ui' http://localhost:3000 'yarn test:ci:ui'
 
   audit:
     name: "Audit"
