+++ packages/arb-token-bridge-ui/package.json
@@ -3,6 +3,7 @@
   "version": "3.3.0",
   "private": true,
   "dependencies": {
+    "@apollo/client": "^3.7.11",
     "@ethersproject/providers": "^5.7.0",
     "@headlessui/react": "^1.7.8",
     "@headlessui/tailwindcss": "^0.1.2",
@@ -14,12 +15,13 @@
     "@tippyjs/react": "^4.2.6",
     "@unstoppabledomains/resolution": "^8.3.3",
     "@walletconnect/web3-provider": "^1.6.5",
-    "axios": "^0.21.4",
+    "axios": "^1.3.4",
     "boring-avatars": "^1.7.0",
     "dayjs": "^1.10.6",
     "ethers": "^5.6.0",
     "fetch-retry": "^5.0.3",
     "framer-motion": "^4.1.17",
+    "graphql": "^16.6.0",
     "lodash-es": "^4.17.21",
     "next": "^13.1.5",
     "next-query-params": "^4.1.0",
