+++ packages/arb-token-bridge-ui/setupTests.ts
@@ -1,0 +1,3 @@
+import 'cross-fetch/polyfill'
+
+jest.setTimeout(25000)
