+++ packages/arb-token-bridge-ui/src/util/deposits/__tests__/fetchDepositsFromSubgraph.test.ts
@@ -1,7 +1,3 @@
-/**
- * @jest-environment jsdom
- */
-
 import {
   getQueryCoveringClassicOnlyWithoutResults,
   getQueryCoveringClassicOnlyWithResults,
