+++ packages/arb-token-bridge-ui/src/util/deposits/fetchDepositsFromSubgraph.ts
@@ -1,4 +1,4 @@
-import { getAPIBaseUrl, sanitizeQueryParams } from './../util'
+import { getAPIBaseUrl, sanitizeQueryParams } from './../index'
 
 export type FetchDepositsFromSubgraphResult = {
   receiver: string
