+++ packages/arb-token-bridge-ui/src/util/index.ts
@@ -44,3 +44,13 @@ export const loadEnvironmentVariableWithFallback = ({
     `
   )
 }
+
+export const sanitizeQueryParams = (data: any) => {
+  return JSON.parse(JSON.stringify(data))
+}
+
+export const getAPIBaseUrl = () => {
+  // if dev environment, eg. tests, then prepend actual running environment
+  // Resolves: next-js-error-only-absolute-urls-are-supported in test:ci:sdk
+  return process.env.NODE_ENV === 'test' ? 'http://localhost:3000' : ''
+}
