+++ packages/token-bridge-sdk/package.json
@@ -27,16 +27,13 @@
     "test:ci": "jest --ci"
   },
   "dependencies": {
-    "@apollo/client": "^3.7.0",
     "@arbitrum/sdk": "^3.1.0",
     "@rehooks/local-storage": "^2.4.4",
     "@uniswap/token-lists": "^1.0.0-beta.27",
     "ajv": "^8.6.3",
     "ajv-formats": "^2.1.1",
-    "axios": "^0.21.4",
     "cross-fetch": "^3.1.5",
     "ethers": "^5.6.0",
-    "graphql": "^16.6.0",
     "swr": "^1.3.0"
   },
   "devDependencies": {
